+++ lib/core/presentation/stores/reit_list_store.dart
@@ -3,7 +3,7 @@ import 'package:fii_app/modules/reit_list/domain/entities/reit_list_sort_option.
 import 'package:mobx/mobx.dart';
 
 import 'package:fii_app/core/domain/entities/reit.dart';
-import '../../domain/usecases/get_all_reits.dart';
+import '../../../modules/reit_list/domain/usecases/get_all_reits.dart';
 
 part 'reit_list_store.g.dart';
