+++ lib/modules/reit_list/presentation/pages/reit_list_page.dart
@@ -11,7 +11,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:get_it/get_it.dart';
 
-import '../stores/reit_list_store.dart';
+import '../../../../core/presentation/stores/reit_list_store.dart';
 
 class ReitListPage extends StatelessWidget {
   final reitListStore = GetIt.I.get<ReitListStore>();
