+++ test/modules/reit_list/presentation/stores/reit_list_store_test.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:fii_app/core/domain/entities/reit.dart';
 import 'package:fii_app/core/errors/failures.dart';
 import 'package:fii_app/modules/reit_list/domain/usecases/get_all_reits.dart';
-import 'package:fii_app/modules/reit_list/presentation/stores/reit_list_store.dart';
+import 'package:fii_app/core/presentation/stores/reit_list_store.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
